Checkpoint_10: Update December 8th

+++ app/controllers/collaborations_controller.rb
@@ -1,7 +1,12 @@
 class CollaborationsController < ApplicationController
+  include Pundit
+
+  before_action :authenticate_user!
+  
   def create
     @wiki = Wiki.find(params[:wiki_id])
     @user = User.find_by_username(params[:search])
+    authorize @collaboration
 
     if @user
       @collaboration = Collaboration.new(wiki_id: params[:wiki_id], user_id: @user.id)
@@ -15,7 +20,7 @@ class CollaborationsController < ApplicationController
       end
     else
       flash[:error] = "The username is invalid. Please try again."
-      render :show 
+      render :show
     end
   end
 
@@ -23,6 +28,7 @@ class CollaborationsController < ApplicationController
   def destroy
   #  @wiki = Wiki.find(params[:wiki_id])
     @collaboration = Collaboration.find(params[:id])
+    authorize @collaboration
 
     if @collaboration.destroy
       flash[:notice] = "Collaborator removed from wiki."
